Add numeric distances and sort linalg.py
Sort builtin classes in linalg.py
Move Numeric Distances to and Distance and Similarity Measures
